backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
openai==1.108.1
SQLAlchemy==2.0.43
psycopg[binary]==3.2.10
pgvector==0.4.1
python-multipart==0.0.20
pypdf==5.9.0
python-docx==1.2.0
httpx==0.28.1
